expense: move Expenses to be stored under Room
Expenses was to be stored in XML under Guest previously.

However, Expenses are supposed to be stored under Room, as Guest does not have an Expenses attribute.

All usage of Expenses in Guest and XmlAdaptedPerson has been shifted to Room and XmlAdaptedRoom.
