--- src/main/java/seedu/address/model/room/Expenses.java
@@ -1,7 +1,0 @@
-package seedu.address.model.room;
-
-/**
- * Dummy class.
- */
-public class Expenses {
-}
