+++ src/main/java/seedu/address/storage/XmlAdaptedPerson.java
@@ -38,9 +38,6 @@ public class XmlAdaptedPerson {
     @XmlElement
     private List<XmlAdaptedTag> tagged = new ArrayList<>();
 
-    @XmlElement
-    private List<XmlAdaptedExpense> expenses = new ArrayList<>();
-
     /**
      * Constructs an XmlAdaptedPerson.
      * This is the no-arg constructor that is required by JAXB.
@@ -50,17 +47,13 @@ public class XmlAdaptedPerson {
     /**
      * Constructs an {@code XmlAdaptedPerson} with the given guest details.
      */
-    public XmlAdaptedPerson(String name, String phone, String email, String address,
-                            List<XmlAdaptedTag> tagged, List<XmlAdaptedExpense> expenses) {
+    public XmlAdaptedPerson(String name, String phone, String email, String address, List<XmlAdaptedTag> tagged) {
         this.name = name;
         this.phone = phone;
         this.email = email;
         this.address = address;
         if (tagged != null) {
             this.tagged = new ArrayList<>(tagged);
-        }
-        if (expenses != null) {
-            this.expenses = expenses;
         }
     }
 
@@ -77,9 +70,6 @@ public class XmlAdaptedPerson {
         tagged = source.getTags().stream()
                 .map(XmlAdaptedTag::new)
                 .collect(Collectors.toList());
-        expenses = source.getExpensesList().stream()
-                .map(XmlAdaptedExpense::new)
-                .collect(Collectors.toList());
     }
 
     /**
@@ -91,11 +81,6 @@ public class XmlAdaptedPerson {
         final List<Tag> personTags = new ArrayList<>();
         for (XmlAdaptedTag tag : tagged) {
             personTags.add(tag.toModelType());
-        }
-
-        final List<Expense> expenseList = new ArrayList<>();
-        for (XmlAdaptedExpense expense : expenses) {
-            expenseList.add(expense.toModelType());
         }
 
         if (name == null) {
@@ -131,8 +116,7 @@ public class XmlAdaptedPerson {
         final Address modelAddress = new Address(address);
 
         final Set<Tag> modelTags = new HashSet<>(personTags);
-        final Expenses expenses = new Expenses(expenseList);
-        return new Guest(modelName, modelPhone, modelEmail, modelAddress, modelTags, expenses);
+        return new Guest(modelName, modelPhone, modelEmail, modelAddress, modelTags);
     }
 
     @Override
